refactor(reset-form): grouped the reset form components and hooks in features/resetForm
