+++ src/features/bottomMenu/components/HistoryRequest/HistoryRequest.jsx
@@ -15,7 +15,7 @@ import {
   getIstexIdsFromIstexIdQueryString,
 } from '@/lib/istexApi';
 import { buildFullIstexDlUrl } from '@/lib/utils';
-import useResetForm from '@/hooks/useResetForm';
+import useResetForm from '@/features/resetForm/hooks/useResetForm';
 import { useEventEmitterContext } from '@/contexts/EventEmitterContext';
 import { useHistoryContext } from '@/contexts/HistoryContext';
