+++ src/features/download/components/DownloadButton/DownloadButton.jsx
@@ -5,7 +5,7 @@ import { Tooltip } from 'flowbite-react';
 import ModalDownloadRewiews from './ModalDownloadRewiews';
 
 import { buildFullApiUrl, sendDownloadApiRequest, sendSaveQIdApiRequest } from '@/lib/istexApi';
-import useResetForm from '@/hooks/useResetForm';
+import useResetForm from '@/features/resetForm/hooks/useResetForm';
 import { useEventEmitterContext } from '@/contexts/EventEmitterContext';
 import { useHistoryContext } from '@/contexts/HistoryContext';
