+++ src/features/query/components/ExamplesButton/Example.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Tooltip } from 'flowbite-react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import useResetForm from '@/hooks/useResetForm';
+import useResetForm from '@/features/resetForm/hooks/useResetForm';
 import { useEventEmitterContext } from '@/contexts/EventEmitterContext';
 
 export default function Example ({ info, closeModal }) {
